[Bronze II] Title: Hashing, Time: 72 ms, Memory: 12168 KB, Score: 50 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.kt
@@ -10,11 +10,11 @@ fun main() {
 
     var total = 0L
     for (i in 0 until N) {
-        var r = 1
+        var r = 1L
         repeat(i) {
-            r *= 31
+            r *= 31L
         }
-        total += (str[i] - 'a' + 1).toLong() * r.toLong()
+        total += (str[i] - 'a' + 1).toLong() * r
     }
 
     println(total % 1234567891)
